+++ app/Http/Controllers/BrandController.php
@@ -65,7 +65,10 @@ class BrandController extends Controller
             $bRand = Brand::findOrFail($id);
 
             if (isset($file)) {
-                Storage::delete($bRand->img_url);
+                if($bRand->img_url){
+                    Storage::delete($bRand->img_url);
+                }
+                
                 $path = $file->store('logo_brand');
                 $bRand->img_url = $path;
             }
@@ -84,9 +87,7 @@ class BrandController extends Controller
         if (empty($bRand)) {
             return "Thương hiệu không tồn tại";
         }
-        if($bRand->img_url){
-            Storage::delete($bRand->img_url);
-        }
+        
         
         $bRand->delete();
         return redirect()->route('brand.list')->with(['Success' => "Xóa thương hiệu {$bRand->name} thành công!"]);
@@ -99,10 +100,4 @@ class BrandController extends Controller
         return redirect()->route('brand.list')->with(['Success' => "Phục hồi thương hiệu {$bRand->name} thành công "]);
     }
 
-    public function deleted($id)
-    {
-        $bRand = Brand::withTrashed()->findOrFail($id);
-        $bRand->forceDelete();
-        return redirect()->route('brand.list')->with(['Success' => "Xóa vĩnh viễn thương hiệu {$bRand->name} thành công "]);
-    }
 }
